bot/requirements.txt
asyncio
aiogram
asyncpg
sqlalchemy
nest-asyncio
python-dotenv